Rename src to vis
