+++ vis/view_config.cpp
