+++ vis/view_config.h
